#12 - Host and join game - fix duplications
