+++ app/src/main/java/com/example/cataniaunited/ui/common/BackgroundScreen.kt
@@ -1,0 +1,27 @@
+package com.example.cataniaunited.ui.common
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import com.example.cataniaunited.R
+
+@Composable
+fun BackgroundScreen(content: @Composable BoxScope.() -> Unit) {
+    Box(modifier = Modifier.fillMaxSize()) {
+        Image(
+            painter = painterResource(id = R.drawable.catan_starting_page_background),
+            contentDescription = "Background",
+            modifier = Modifier.fillMaxSize(),
+            contentScale = ContentScale.Crop
+        )
+        Box( modifier = Modifier.align(Alignment.Center)){
+            content()
+        }
+    }
+}
