+++ app/src/main/java/com/example/cataniaunited/ui/common/OrDivider.kt
@@ -1,0 +1,18 @@
+package com.example.cataniaunited.ui.common
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun OrDivider() {
+    Text(
+        text = "OR",
+        style = MaterialTheme.typography.titleLarge,
+        color = MaterialTheme.colorScheme.onPrimary,
+        modifier = Modifier.padding(top = 8.dp, bottom = 8.dp)
+    )
+}
